Restructure project into a well formed Python package
- Add pyproject.toml and uv.lock
- Move sources into src/challenge_toolkit
- Expose CLI as 'ctf-toolkit' via project.scripts and set uv_build
  backend
- Remove requirements.txt and legacy test runner
- Modify imports to always start from package name
