+++ pyproject.toml
@@ -1,0 +1,23 @@
+[project]
+name = "challenge-toolkit"
+version = "0.1.0"
+description = "CTFp challenge development toolkit"
+readme = "README.md"
+requires-python = ">=3.14"
+license-files = ["LICENSE"]
+dependencies = [
+  "python-slugify==8.0.4",
+  "pyyaml==6.0.2",
+]
+
+[project.scripts]
+ctf-toolkit = "challenge_toolkit.ctf:main"
+
+[dependency-groups]
+dev = [
+  "pytest>=9.0.2",
+]
+
+[build-system]
+requires = ["uv_build>=0.9.28,<0.10.0"]
+build-backend = "uv_build"
