+++ src/challenge_toolkit/commands/challenge_creator.py
@@ -7,22 +7,22 @@ Prompts the user for inputs and generates a template for a CTF challenge.
 import sys
 import argparse
 
-from library.config import CHALL_TYPES, DIFFICULTIES, FLAG_FORMAT, INSTANCED_TYPES, CATEGORIES
-from library.utils import Utils
-from library.data import Challenge, DockerfileLocation
-from library.generator import Generator as OSGenerator
+from challenge_toolkit.library.config import CHALL_TYPES, DIFFICULTIES, FLAG_FORMAT, INSTANCED_TYPES, CATEGORIES
+from challenge_toolkit.library.utils import Utils
+from challenge_toolkit.library.data import Challenge, DockerfileLocation
+from challenge_toolkit.library.generator import Generator as OSGenerator
 
-class Args:    
+class Args:
     args = None
     subcommand = False
-    
+
     def __init__(self, parent_parser = None):
         if parent_parser:
             self.subcommand = True
             self.parser = parent_parser.add_parser("create", help="Template Generator for CTF Challenges")
         else:
             self.parser = argparse.ArgumentParser(description="Template Generator for CTF Challenges")
-        
+
         self.parser.add_argument("--no-prompts", help="Skip prompts and use default values", action="store_true")
         self.parser.add_argument("--name", help="Name of the challenge")
         self.parser.add_argument("--slug", help="Slug of the challenge")
@@ -39,16 +39,16 @@ class Args:
         self.parser.add_argument("--dockerfile-context", help="Context of the Dockerfile", default="src/")
         self.parser.add_argument("--dockerfile-identifier", help="Identifier of the Dockerfile", default=None)
         self.parser.add_argument("--handout_location", help="Location of the handout", default="handout")
-        
+
     def parse(self):
         if self.subcommand:
             self.args = self.parser.parse_args(sys.argv[2:])
         else:
             self.args = self.parser.parse_args()
-        
+
     def prompt(self, challenge: Challenge):
         args = self.args
-        
+
         if args is None:
             # Convert to object if args is None
             args = self.args = self.parser.parse_args()
@@ -63,7 +63,7 @@ class Args:
                     print("Invalid name. Please try again.")
         else:
             challenge.set_name(args.name)
-        
+
         if args.slug is None:
             while True:
                 try:
@@ -73,7 +73,7 @@ class Args:
                     print("Invalid slug. Please try again.")
         else:
             challenge.set_slug(args.slug)
-        
+
         if args.author is None:
             while True:
                 try:
@@ -83,7 +83,7 @@ class Args:
                     print("Invalid author. Please try again.")
         else:
             challenge.set_author(args.author)
-                
+
         if args.category is None:
             while True:
                 try:
@@ -93,7 +93,7 @@ class Args:
                     print("Invalid category. Please try again.")
         else:
             challenge.set_category(args.category)
-                
+
         if args.difficulty is None:
             while True:
                 try:
@@ -103,7 +103,7 @@ class Args:
                     print("Invalid difficulty. Please try again.")
         else:
             challenge.set_difficulty(args.difficulty)
-        
+
         prompted_type = None
         if args.type is None:
             while True:
@@ -116,7 +116,7 @@ class Args:
         else:
             challenge.set_type(args.type)
             prompted_type = args.type
-                
+
         if args.flag is None:
             while True:
                 try:
@@ -126,7 +126,7 @@ class Args:
                     print("Invalid flag. Please try again.")
         else:
             challenge.set_flag(args.flag)
-                
+
         if args.points is None:
             while True:
                 try:
@@ -136,7 +136,7 @@ class Args:
                     print("Invalid points. Please try again.")
         else:
             challenge.set_points(args.points)
-                
+
         if args.min_points is None:
             while True:
                 try:
@@ -146,7 +146,7 @@ class Args:
                     print("Invalid minimum points. Please try again.")
         else:
             challenge.set_min_points(args.min_points)
-        
+
         if (args.type in [ "instanced", "shared" ] or prompted_type in [ "instanced", "shared" ]) and args.instanced_type == "none":
             while True:
                 try:
@@ -158,7 +158,7 @@ class Args:
             challenge.set_instanced_type(args.instanced_type)
         else:
             challenge.set_instanced_type("none")
-                
+
         if args.description_location == "description.md":
             while True:
                 try:
@@ -168,7 +168,7 @@ class Args:
                     print("Invalid description location. Please try again.")
         else:
             challenge.set_description_location(args.description_location)
-                
+
         if args.dockerfile_location is None or args.dockerfile_location == "src/Dockerfile":
             contains_docker = input("Does the challenge contain a Dockerfile? (y/N): ").lower() == "y"
             if contains_docker:
@@ -177,7 +177,7 @@ class Args:
                         dockerfile_location = input("Location of the Dockerfile (src/Dockerfile): ") or "src/Dockerfile"
                         dockerfile_context = input("Context of the Dockerfile (src/): ") or "src/"
                         dockerfile_identifier = input("Identifier of the Dockerfile: ") or None
-                        
+
                         challenge.add_dockerfile_location([ DockerfileLocation(dockerfile_location, dockerfile_context, dockerfile_identifier) ])
                         break
                     except ValueError:
@@ -197,7 +197,7 @@ class Generator:
         self.challenge = challenge
         self.path = Utils.get_challenge_dir(challenge.category, challenge.slug)
         self.generator = OSGenerator(challenge)
-    
+
     def generate(self):
         self.generator.build()
 
@@ -205,12 +205,12 @@ class ChallengeCreator:
     args = None
     parent_parser = None
 
-    def __init__(self, parent_parser = None):   
+    def __init__(self, parent_parser = None):
         self.parent_parser = parent_parser
-  
+
     def register_subcommand(self):
         self.args = Args(self.parent_parser)
-  
+
     def run(self):
         if not self.args:
             arguments = Args(self.parent_parser)
@@ -218,48 +218,48 @@ class ChallengeCreator:
             self.args = arguments
         else:
             self.args.parse()
-        
+
         arguments = self.args
         args = self.args.args
-        
+
         if not args:
             print("Error parsing arguments. Please run with --help to see available options.")
             sys.exit(1)
-        
+
         if args.name and args.slug is None:
             args.slug = Utils.slugify(args.name) if args.name else "challenge"
-        
+
         challenge = None
         if args.no_prompts == False:
             challenge = Challenge(name="demo", slug="demo", author="demo", category="misc", difficulty="easy", type="static", flag="flag{demo_flag}")
             arguments.prompt(challenge)
 
             print("\nInformation filled out.")
-            
+
             print("\nInformation for the challenge:")
             print(challenge)
-            
+
             print("\nIs the information correct?")
             if (input("Y/n: ") or "y").lower() != "y":
                 print("Exiting...")
                 sys.exit(1)
-        
+
         else:
             challenge = Challenge(
-                name = args.name, 
-                slug = args.slug, 
-                author = args.author, 
-                category = args.category, 
-                difficulty = args.difficulty, 
+                name = args.name,
+                slug = args.slug,
+                author = args.author,
+                category = args.category,
+                difficulty = args.difficulty,
                 type = args.type,
-                instanced_type = args.instanced_type or "none", 
-                flag = args.flag, 
-                points = args.points or 1000, 
+                instanced_type = args.instanced_type or "none",
+                flag = args.flag,
+                points = args.points or 1000,
                 min_points = args.min_points or 100,
                 description_location = args.description_location,
                 handout_dir = args.handout_location
             )
-        
+
             if args.type != "static":
                 try:
                     if args.dockerfile_location:
@@ -269,8 +269,7 @@ class ChallengeCreator:
 
         generator = Generator(challenge)
         generator.generate()
-        
+
 
 if __name__ == '__main__':
     ChallengeCreator().run()
-
