+++ src/challenge_toolkit/commands/slugify.py
@@ -1,13 +1,13 @@
 import sys
 import argparse
 
-from library.utils import Utils
+from challenge_toolkit.library.utils import Utils
 
 class Args:
     args = None
     slug: str = ""
     subcommand = False
-    
+
     def __init__(self, parent_parser = None):
         if parent_parser:
             self.subcommand = True
@@ -16,19 +16,19 @@ class Args:
             self.parser = argparse.ArgumentParser(description="Slugify a string for use in challenge slug")
 
         self.parser.add_argument("name", help="Name to slugify")
-    
+
     def parse(self):
         if self.subcommand:
             self.args = self.parser.parse_args(sys.argv[2:])
         else:
             self.args = self.parser.parse_args()
-        
+
         if not self.args.name:
             print("Name is required")
             sys.exit(1)
 
         self.slug = self.args.name
-        
+
 class Slugify:
     @staticmethod
     def run(name: str) -> str:
@@ -43,10 +43,10 @@ class SlugifyCommand:
 
     def __init__(self, parent_parser = None):
         self.parent_parser = parent_parser
-  
+
     def register_subcommand(self):
         self.args = Args(self.parent_parser)
-  
+
     def run(self):
         if not self.args:
             arguments = Args(self.parent_parser)
@@ -54,8 +54,7 @@ class SlugifyCommand:
             self.args = arguments
         else:
             self.args.parse()
-        
+
         args = self.args
 
         print(Slugify.run(args.slug))
-
