+++ src/challenge_toolkit/tests/data/full-example-multi-flag-object.yml
