+++ src/challenge_toolkit/tests/data/full-example-multi-flag.json
