+++ src/challenge_toolkit/tests/data/full-example.yml
