+++ src/challenge_toolkit/tests/data/minimal-example.yml
