--- src/requirements.txt
@@ -1,2 +1,0 @@
-pyyaml==6.0.2
-python-slugify==8.0.4
